Implement changes requested
