+++ src/main/java/seedu/address/logic/commands/SortContactCommand.java
@@ -7,6 +7,7 @@ import java.util.Comparator;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.comparators.PersonComparator;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
 
@@ -30,37 +31,6 @@ public class SortContactCommand extends Command {
 
     public static final String MESSAGE_ARGUMENTS = "Index: %1$d";
 
-    private static final Comparator<Person> NAME_COMPARATOR = new Comparator<Person>() {
-        @Override
-        public int compare(Person o1, Person o2) {
-            return o1.getName().fullName.compareToIgnoreCase(o2.getName().fullName);
-        }
-    };
-
-    private static final Comparator<Person> ADDRESS_COMPARATOR = new Comparator<Person>() {
-        @Override
-        public int compare(Person o1, Person o2) {
-            return o1.getAddress().value.compareToIgnoreCase(o2.getAddress().value);
-        }
-    };
-
-    private static final Comparator<Person> TAG_COMPARATOR = new Comparator<Person>() {
-        @Override
-        public int compare(Person o1, Person o2) {
-            if (o1.getTags().size() == 0 && o2.getTags().size() != 0) {
-                return 1;
-            }
-            if (o1.getTags().size() != 0 && o2.getTags().size() == 0) {
-                return -1;
-            }
-            if (o1.getTags().size() == 0 && o2.getTags().size() == 0) {
-                return o1.getName().fullName.compareToIgnoreCase(o2.getName().fullName);
-            }
-            return o1.getTags().iterator().next().tagName
-                    .compareToIgnoreCase(o2.getTags().iterator().next().tagName);
-        }
-    };
-
     private final Index index;
 
     /**
@@ -74,7 +44,7 @@ public class SortContactCommand extends Command {
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
-        model.sortPerson(chooseComparator(index));
+        model.sortPerson(PersonComparator.chooseComparator(index));
 
         return new CommandResult(indexMessage(index));
     }
@@ -94,21 +64,6 @@ public class SortContactCommand extends Command {
         default:
             return "Invalid index entered, refer to below for the command's proper usage: "
                     + MESSAGE_USAGE;
-        }
-    }
-
-    /**
-     * Returns the appropriate comparator for the model manager to sort with
-     */
-    public Comparator<Person> chooseComparator(Index index) {
-        int input = index.getOneBased();
-        switch (input) {
-        case 2:
-            return ADDRESS_COMPARATOR;
-        case 3:
-            return TAG_COMPARATOR;
-        default:
-            return NAME_COMPARATOR;
         }
     }
