+++ src/main/java/seedu/address/logic/comparators/PersonComparator.java
@@ -1,0 +1,58 @@
+package seedu.address.logic.comparators;
+
+import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.person.Person;
+
+import java.util.Comparator;
+
+public class PersonComparator {
+
+    private static final Comparator<Person> NAME_COMPARATOR = new Comparator<Person>() {
+        @Override
+        public int compare(Person o1, Person o2) {
+            return o1.getName().fullName.compareToIgnoreCase(o2.getName().fullName);
+        }
+    };
+
+    private static final Comparator<Person> ADDRESS_COMPARATOR = new Comparator<Person>() {
+        @Override
+        public int compare(Person o1, Person o2) {
+            return o1.getAddress().value.compareToIgnoreCase(o2.getAddress().value);
+        }
+    };
+
+    private static final Comparator<Person> TAG_COMPARATOR = new Comparator<Person>() {
+        @Override
+        public int compare(Person o1, Person o2) {
+            if (o1.getTags().size() == 0 && o2.getTags().size() != 0) {
+                return 1;
+            }
+            if (o1.getTags().size() != 0 && o2.getTags().size() == 0) {
+                return -1;
+            }
+            if (o1.getTags().size() == 0 && o2.getTags().size() == 0) {
+                return o1.getName().fullName.compareToIgnoreCase(o2.getName().fullName);
+            }
+            return o1.getTags().iterator().next().tagName
+                    .compareToIgnoreCase(o2.getTags().iterator().next().tagName);
+        }
+    };
+
+    /**
+     * Returns the appropriate comparator for the model manager to sort with
+     */
+    public static Comparator<Person> chooseComparator(Index index) throws CommandException {
+        int input = index.getOneBased();
+        switch (input) {
+            case 1:
+                return NAME_COMPARATOR;
+            case 2:
+                return ADDRESS_COMPARATOR;
+            case 3:
+                return TAG_COMPARATOR;
+            default:
+                throw new CommandException("Index should be between 1 to 3");
+        }
+    }
+}
